pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "Local job search assistant with a FastAPI backend and React frontend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "feedparser>=6.0.12",
    "langchain-core>=1.2.16",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.9",
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "streamlit>=1.54.0",
    "tavily-python>=0.7.22",
    "uvicorn>=0.35.0",
]
